Applied the cocomed changes

+++ src/xrGame/Level_network_spawn.cpp
@@ -84,7 +84,7 @@ void CLevel::g_cl_Spawn		(LPCSTR name, u8 rp, u16 flags, Fvector pos)
 void CLevel::g_sv_Spawn		(CSE_Abstract* E)
 {
 #ifdef DEBUG_MEMORY_MANAGER
-	u32							E_mem = 0;
+	size_t							E_mem = 0;
 	if (g_bMEMO)	{
 		lua_gc					(ai().script_engine().lua(),LUA_GCCOLLECT,0);
 		lua_gc					(ai().script_engine().lua(),LUA_GCCOLLECT,0);
@@ -192,7 +192,7 @@ void CLevel::g_sv_Spawn		(CSE_Abstract* E)
 	if (g_bMEMO) {
 		lua_gc					(ai().script_engine().lua(),LUA_GCCOLLECT,0);
 		lua_gc					(ai().script_engine().lua(),LUA_GCCOLLECT,0);
-		Msg						("* %20s : %d bytes, %d ops", *E->s_name,Memory.mem_usage()-E_mem, Memory.stat_calls );
+		Msg						("* %20s : %lld bytes, %d ops", *E->s_name,Memory.mem_usage()-E_mem, Memory.stat_calls );
 	}
 #endif // DEBUG_MEMORY_MANAGER
 }
